refactor: add private methods
